add Layout component
