+++ client/src/containers/Root.js
@@ -11,7 +11,7 @@ import ContactForm from 'components/ContactForm'
 import Dashboard from 'components/Dashboard'
 import withTracker from 'containers/withTracker'
 import Web3, {withNetwork} from 'containers/Web3'
-
+import Layout from 'components/Layout'
 const history = createHistory()
 
 const contactFormTransition = {
@@ -45,27 +45,29 @@ export default class Root extends Component {
           <div>
             <Web3 />
             <div style={{height: '100%'}}>
-              <Route path='/' component={withTracker(withNetwork(App))} />
-              <div className='contact-form-wrapper'>
-                <AnimatedRoute
-                  path='/view/contact-us'
-                  component={withTracker(ContactForm)}
+              <Layout>
+                <Route exact path='/' component={withTracker(withNetwork(App))} />
+                <div className='contact-form-wrapper'>
+                  <AnimatedRoute
+                    path='/view/contact-us'
+                    component={withTracker(ContactForm)}
+                    mapStyles={mapStylesContact}
+                    {...contactFormTransition}
+                  />
+                </div>
+                <Route
+                  path='/view/issuance'
+                  component={withTracker(withNetwork(IssuanceWizard))}
                   mapStyles={mapStylesContact}
                   {...contactFormTransition}
                 />
-              </div>
-              <Route
-                path='/view/issuance'
-                component={withTracker(withNetwork(IssuanceWizard))}
-                mapStyles={mapStylesContact}
-                {...contactFormTransition}
-              />
-              <Route
-                path='/view/dashboard/:address'
-                component={withTracker(withNetwork(Dashboard))}
-                mapStyles={mapStylesContact}
-                {...contactFormTransition}
-              />
+                <Route
+                  path='/view/dashboard/:address'
+                  component={withTracker(withNetwork(Dashboard))}
+                  mapStyles={mapStylesContact}
+                  {...contactFormTransition}
+                />
+              </Layout>
             </div>
           </div>
         </ConnectedRouter>
